% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{abort}
\alias{abort}
\title{Aborts the specified job.}
\usage{
abort(job, ...)
}
\arguments{
\item{job}{The job to abort.}

\item{...}{Extra parameters passed to \code{httr}}
}
\description{
Aborts the specified job.
}

