% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{copyJob}
\alias{copyJob}
\title{Creates a new job by copying an existing job.}
\usage{
copyJob(job, shallow = FALSE, override_data = "", ...)
}
\arguments{
\item{job}{The job.}

\item{shallow}{Indicates if the copy is shallow.}

\item{override_data}{Data to override as a JSON formatted string.}

\item{...}{parameters to override}
}
\description{
The existing job must not be running or waiting for execution. All creation
data are copied over to the new job. By default, the input attachments and
their contents are copied in the new job. If a shallow copy is requested the
new attachment will point to the existing job, and if it is deleted accessing
the attachment wil raise an exception. Output attachments are not copied.
Optionally, a job creation data can be passed to override the parameters and
declare additional or replacement input attachments.
}

