% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{create}
\alias{create}
\title{Create job and submit job.}
\usage{
create(job, ...)
}
\arguments{
\item{job}{The job.}

\item{...}{Extra parameters passed to \code{httr}}
}
\description{
This method creates a job, upload attachments and submit a request to execute the job.
}

