% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{delete}
\alias{delete}
\title{Delete the job.}
\usage{
delete(job, ...)
}
\arguments{
\item{job}{The job.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
the job.
}
\description{
Delete the job.
}

