% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{deleteAllJobs}
\alias{deleteAllJobs}
\title{Delete all jobs for the current user.}
\usage{
deleteAllJobs(client, ...)
}
\arguments{
\item{client}{The client.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
The response from the server.
}
\description{
Delete all jobs for the current user.
}

