% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\docType{class}
\name{DOcplexcloudAttachment-class}
\alias{DOcplexcloudAttachment}
\alias{DOcplexcloudAttachment-class}
\title{The base class for attachments.}
\arguments{
\item{name}{The name of the attachment. If a \code{file} is specified,
the name can be ommited and the basename of the \code{file} is used.}

\item{file}{The name of the file containing the data for the attachment.}

\item{data}{The raw data for the attachment.}
}
\description{
The base class for attachments.
}
\section{Methods}{

\describe{
\item{\code{getName()}}{Returns a name for the attachment.

The name of the attachment is the \code{name} field of the attachement.
If that field is empty or NULL, then the name is the basename of
\code{attachment$file}}
}}
\examples{
# Creates an attachment which content is the content of this file,
# and which name is "model.lp"
a <- DOcplexcloudAttachment$new(file = "/home/joe/models/model.lp")

# Creates an attachment which content is the specified file, and
# which name is "model.lp"
a <- DOcplexcloudAttachment$new(name = "model.lp",
                                file = "/home/joe/models/model_1231.lp")

# create an attachment which data is stored in memory
model <- "Minimize
            obj: x + y
          Subject To
          Bounds
            3 <= x <= 17
            2 <= y
          End"
a <- DOcplexcloudAttachment$new(name="model.lp",
                                data=charToRaw(model))
}
\keyword{internal}

