% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{DOcplexcloudJob}
\alias{DOcplexcloudJob}
\title{The job class. This class provides the API to manage operations on jobs.}
\usage{
DOcplexcloudJob(client = NULL, joburl = NULL, jobid = NULL,
  executionStatus = NULL, attachments = list())
}
\description{
The job class. This class provides the API to manage operations on jobs.
}
\section{Fields}{

\describe{
\item{\code{client}}{The \code{DOcplexcloudClient}.}

\item{\code{joburl}}{The url of the job.}

\item{\code{jobid}}{The id of the job.}

\item{\code{exectutionStatus}}{The execution status if the job has been executed.}

\item{\code{attachments}}{The attachments}
}}
\note{
You normally don't use this to actually create a job, but
  instead you would create a new job from a \code{DOcplexcloudClient}
}
\examples{
\dontrun{
client <- DOcplexcloudClient(url='Your DOcplexcloud base URL',
                             key='Your DOcplexcloud api key')

# create job
job <- submitJob(client, addAttachment(file="model.lp"))
status <- waitForCompletion(job)
}
}

