% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\docType{data}
\name{DOcplexcloudJobResponse}
\alias{DOcplexcloudJobResponse}
\title{The job response class. This class is returned by client methods, storing
job information.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DOcplexcloudJobResponse
}
\description{
The job response class. This class is returned by client methods, storing
job information.
}
\keyword{datasets}

