% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{execute}
\alias{execute}
\title{Submits an execution request for a job.}
\usage{
execute(job, ...)
}
\arguments{
\item{job}{The job.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
The job.
}
\description{
Submits an execution request for a job.
}

