% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{getAllJobs}
\alias{getAllJobs}
\title{Returns all jobs for the current user.}
\usage{
getAllJobs(client, ...)
}
\arguments{
\item{client}{The client.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
a list of DOcplexcloudJob
}
\description{
Returns all jobs for the current user.
}

