% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{getAttachment}
\alias{getAttachment}
\title{Downloads the specified attachment for a job.}
\usage{
getAttachment(job, name, ...)
}
\arguments{
\item{job}{The job.}

\item{name}{The name of the attachment.}

\item{...}{Extra parameters passed to \code{httr}}

\item{convert}{If TRUE, the method try to convert the attachment to a data
structure ready for use (example: parse JSON).}
}
\description{
Attachments are always returned as raw data, unless
the attachment ends with ".json". In that case, the
json is downloaded, parsed and converted from JSON,
unless \code{convert} is false
}

