% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{getAttachmentName}
\alias{getAttachmentName}
\title{Returns a name for the attachment.}
\usage{
getAttachmentName(attachment)
}
\arguments{
\item{attachment}{The attachment}
}
\value{
A name for the attachment.
}
\description{
The name of the attachment is the \code{name} field of the attachement.
If that field is empty or NULL, then the name is the basename of
\code{attachment$file}
}

