% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{getInfo}
\alias{getInfo}
\title{Returns job info for the specified job.}
\usage{
getInfo(job, ...)
}
\arguments{
\item{job}{The job.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
the job info.
}
\description{
Returns job info for the specified job.
}

