% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{getJobUrl}
\alias{getJobUrl}
\title{Returns the joburl if a DOcplexcloudJob is passed. Otherwise return the arg.}
\usage{
getJobUrl(job_or_url)
}
\description{
Returns the joburl if a DOcplexcloudJob is passed. Otherwise return the arg.
}
\keyword{internal}
