% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{getLogs}
\alias{getLogs}
\title{Downloads the specified log for a job.}
\usage{
getLogs(job, ...)
}
\arguments{
\item{job}{The job.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
The logs as a raw data
}
\description{
Downloads the specified log for a job.
}

