% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{makeRequest}
\alias{makeRequest}
\title{Submits a REST request to DOcplexcloud.}
\usage{
makeRequest(client, verb, url, fail_message, ...)
}
\arguments{
\item{client}{The client.}

\item{verb}{The request verb.}

\item{url}{The URL.}

\item{fail_message}{The message to display incase of failure}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
The response.
}
\description{
This makes a request to DOcplexcloud, with the relevant headers.
}
\keyword{internal}

