% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{setAttachment}
\alias{setAttachment}
\title{Attach the specified file or data to the job.
The attachement is just attached. Data is actually uploaded when job is
created.}
\usage{
setAttachment(job, name = NULL, file = NULL, data = NULL)
}
\arguments{
\item{job}{The job}

\item{name}{The name of the attachment. If a \code{file} is specified,
the name can be ommited and the basename of the \code{file} is used.}

\item{file}{The name of the file containing the data for the attachment.}

\item{data}{The raw data for the attachment.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
The job
}
\description{
Attach the specified file or data to the job.
The attachement is just attached. Data is actually uploaded when job is
created.
}

