% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{status}
\alias{status}
\title{Returns the execution status of a job.}
\usage{
status(job, ...)
}
\arguments{
\item{job}{The job.}

\item{...}{Extra parameters passed to \code{httr}}
}
\description{
Returns the execution status of a job.
}

