% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{submitJob}
\alias{submitJob}
\title{Submit a job execution.}
\usage{
submitJob(client, ...)
}
\arguments{
\item{client}{A DOcplexcloudClient.}

\item{...}{Further parameters like attachments.}
}
\value{
A DOcplexcloudJob.
}
\description{
Once the job is executed, it can be monitored using waitForCompletion() then
deleted with delete()
}
\examples{
\dontrun{
# create and execute a job to solve model.lp
job <- submit(client, addAttachment(file="model.lp"))

# create and execute a job which model is stored
# in memory
model <- "Minimize
            obj: x + y
          Subject To
          Bounds
            3 <= x <= 17
            2 <= y
          End"
job <- submit(client, addAttachment(name="model.lp",
                                    data=charToRaw(model)))
}
}

