% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{uploadAttachment}
\alias{uploadAttachment}
\title{Uploads the specified attachment for a job. The data is actually uploaded.
The job must have been created.}
\usage{
uploadAttachment(job, attachment = NULL, ...)
}
\arguments{
\item{job}{The job.}

\item{attachment}{The attachment.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
The job with updated attachments.
}
\description{
Uploads the specified attachment for a job. The data is actually uploaded.
The job must have been created.
}

