% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{waitForCompletion}
\alias{waitForCompletion}
\title{Wait for the specified job to complete.}
\usage{
waitForCompletion(job, waittime = Inf, ...)
}
\arguments{
\item{job}{The job}

\item{waittime}{The maximum time to wait. This default to \code{Inf} if not specified.}

\item{...}{Extra parameters passed to \code{httr}}
}
\value{
the status as a string value.
}
\description{
Wait for the specified job to complete.
}

