% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{addAttachment}
\alias{addAttachment}
\title{Creates a new attachment.}
\usage{
addAttachment(name = NULL, file = NULL, data = NULL)
}
\arguments{
\item{name}{The name of the attachment. If a \code{file} is specified,
the name can be ommited and the basename of the \code{file} is used.}

\item{file}{The name of the file containing the data for the attachment.}

\item{data}{The raw data for the attachment.}
}
\value{
an instance of DOcplexcloudAttachment with the specified attachment data.
}
\description{
Creates a new attachment.
}
\examples{
# Creates an attachment which content is the content of this file,
# and which name is "model.lp"
a <- addAttachment(file = "/home/joe/models/model.lp")

# Creates an attachment which content is the specified file, and
# which name is "model.lp"
a <- addAttachment(name = "model.lp",
                   file = "/home/joe/models/model_1231.lp")

# create an attachment which data is stored in memory
model <- "Minimize
            obj: x + y
          Subject To
          Bounds
            3 <= x <= 17
            2 <= y
          End"
a <- addAttachment(name="model.lp",
                   data=charToRaw(model))
}

