% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{addAttachment}
\alias{addAttachment}
\title{Creates a new attachment.}
\usage{
addAttachment(name = NULL, file = NULL, data = NULL)
}
\arguments{
\item{name}{The name of the attachment. If a \code{file} is specified,
the name can be ommited and the basename of the \code{file} is used.
If \code{name} is a string pointing to a file, and \code{file} is not
specified, the content of the file pointed by \code{name} is used as
attachment data.}

\item{file}{The name of the file containing the data for the attachment.}

\item{data}{The data for the attachment. If \code{data} is a list which
\code{names(data)} is not empty, the data is first converted to csv.}
}
\value{
An attachment object suitable to use with methods of
   DOcplexcloudClient and DOcplexcloudJob
}
\description{
Creates a new attachment.
}
\examples{
# Creates an attachment which content is the content of this file,
# and which name is "model.lp" (the 3 expressions are equivalent):
a <- addAttachment(file = "/home/joe/models/model.lp")
a <- addAttachment(name = "/home/joe/models/model.lp")
a <- addAttachment("/home/joe/models/model.lp")

# Creates an attachment which content is the specified file, and
# which name is "model.lp"
a <- addAttachment(name = "model.lp",
                   file = "/home/joe/models/model_1231.lp")

# create an attachment which data is stored in memory
model <- "Minimize
            obj: x + y
          Subject To
          Bounds
            3 <= x <= 17
            2 <= y
          End"
a <- addAttachment(name="model.lp",
                   data=charToRaw(model))
\dontrun{
# submits and run the specified model.lp:
job <- client$submitJob(addAttachment("/home/joe/models/model.lp"))
}
}

