% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docplexcloud.R
\docType{package}
\name{docplexcloud}
\alias{docplexcloud}
\alias{docplexcloud-package}
\title{docplexcloud: The R client for Decision Optimization on cloud}
\description{
docplexcloud: The R client for Decision Optimization on cloud
}

