% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\docType{data}
\name{DOcplexcloudAttachment}
\alias{DOcplexcloudAttachment}
\title{The base class for attachments.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DOcplexcloudAttachment
}
\description{
The base class for attachments.
}
\section{Fields}{

\describe{
\item{\code{name}}{The name of the attachment. If a \code{file} is specified,
the name can be ommited and the basename of the \code{file} is used.}

\item{\code{file}}{The name of the file containing the data for the attachment.}

\item{\code{data}}{The raw data for the attachment.}
}}
\keyword{internal}

