% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{DOcplexcloudClient}
\alias{DOcplexcloudClient}
\title{The DOcplexcloud client class.}
\format{The DOcplexcloud client class.}
\usage{
DOcplexcloudClient
}
\description{
The client makes it easier to work with the
solve service by encapsulating connection parameters (connection URL, api key)
and provides support for features of DOcplexcloud solve service not tied to a
particular job, like listing
all jobs for the current user or deleting all jobs.
}
\section{Fields}{

\describe{
\item{\code{url}}{The DOcplexcloud URL.}

\item{\code{key}}{The DOcplexcloud api key.}

\item{\code{verbose}}{If TRUE, sets the verbose mode.}
}}
\section{Methods}{

\describe{
  \item{\code{new(url, key, verbose=FALSE)}}{Initialize a new client with
      the specified \code{url} and \code{key} parameters.

      If \code{url} is not specified, the value of environment variable
      \code{DOCPLEXCLOUD_URL} is used. If \code{key} is not specified, the
      value of environment variable \code{key} is used.

      When \code{verbose} is TRUE, extra information is printed when the
      client API is called.
  }
  \item{\code{submitJob(..., wait=TRUE)}}{Submits a job execution.

      This method creates the job, upoads the attachments, submit an execution
      request, then wait for completion of the job, unless \code{wait} is
      \code{FALSE}.

      Returns a \code{DOcplexcloudJob}.
  }
  \item{\code{getAllJobs(...)}}{Returns all jobs for the current user.}
  \item{\code{deleteAllJobs(...)}}{Delete all jobs for the current user.}
  \item{\code{waitForCompletion(job, waittime=Inf, ...)}}{
      Wait for the specified \code{job} to complete.

      \code{job} can be either a \code{DOcplexcloudJob} or a job URL.

      \code{waittime} The maximum time to wait. This default to \code{Inf}
      if not specified.

      Returns the job execution status, which can be: \code{CREATED},
      \code{NOT_STARTED}, \code{RUNNING}, \code{INTERRUPTING},
      \code{INTERRUPTED}, \code{FAILED}, \code{PROCESSED}
  }
}
}
\examples{
\dontrun{
# create a new client, using environment variables for url and api key
client <- DOcplexCloudClient$new()

# create and execute a job to solve model.lp
job <- client$submitJob(addAttachment(file="model.lp"))

# create and execute a job which model is stored
# in memory
model <- "Minimize
            obj: x + y
          Subject To
          Bounds
            3 <= x <= 17
            2 <= y
          End"
job <- client$submitJob(addAttachment(name="model.lp",
                                      data=charToRaw(model)))
# download solution
solution = client$getAttachment(job, "solution.json")
}
Creates a job.

}
\keyword{datasets}

