% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{DOcplexcloudClient}
\alias{DOcplexcloudClient}
\title{Creates a new instance of \code{DOcplexcloudClient}}
\usage{
DOcplexcloudClient(url = NULL, key = NULL, verbose = FALSE)
}
\arguments{
\item{url}{The DOcplexcloud url. If not specified, the value of environment
variable DOCPLEXCLOUD_URL is used.}

\item{key}{The DOcplexcloud api key. If not specified, the value of environment
variable DOCPLEXCLOUD_KEY is used}

\item{verbose}{If TRUE, sets the verbose mode.}
}
\value{
The DOcplexcloudClient
}
\description{
Creates a new instance of \code{DOcplexcloudClient}
}

