% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\docType{data}
\name{DOcplexcloudJob}
\alias{DOcplexcloudJob}
\title{The job class.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DOcplexcloudJob
}
\description{
This class is returned by client methods such as
\code{DOcplexcloudClient.submitJob()} and is used to store information for jobs
created and executed on DOcplexcloud.
}
\section{Fields}{

\describe{
\item{\code{joburl}}{The URL of the job.}

\item{\code{jobid}}{The id of the job.}

\item{\code{exectutionStatus}}{The execution status if the job has been executed.}
}}

\examples{
\dontrun{
client <- DOcplexcloudClient(url='Your DOcplexcloud base URL',
                             key='Your DOcplexcloud api key')

# Create job and wait for completion
job <- client$submitJob(addAttachment(file="model.lp"))
# The job status
status <- job$executionStatus
}
}
\keyword{datasets}
