% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{DOcplexcloudJob}
\alias{DOcplexcloudJob}
\title{Creates a new instance of \code{DOcplexcloudJob}}
\usage{
DOcplexcloudJob(client = NULL, joburl = NULL, attachments = list())
}
\arguments{
\item{client}{The \code{DOcplexcloudClient}.}

\item{joburl}{The url of the job.}

\item{attachments}{The attachments}
}
\description{
Creates a new instance of \code{DOcplexcloudJob}
}
\note{
You normally don't use this to actually create a job, but
  instead you would create a new job from a \code{DOcplexcloudClient}
}
\examples{
\dontrun{
client <- DOcplexcloudClient(url='Your DOcplexcloud base URL',
                             key='Your DOcplexcloud api key')

# create job
job <- submitJob(client, addAttachment(file="model.lp"))
status <- waitForCompletion(job)
}
}
\keyword{internal}

